Refactoring and Physics
- Duplicate object (Ctrl + D)
- Delete object (Delete)
- Edit Z-Index on runtime
- Fixing the fvcking halfsize
- Adding Enum dropdown support for ImGui

+++ mario/src/main/java/components/GridLines.java
@@ -15,14 +15,14 @@ public class GridLines extends Component{
         Vector2f cameraPos = camera.position;
         Vector2f projectionSize = camera.getProjectionSize();
 
-        int firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
-        int firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
+        float firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
+        float firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
 
         int numVtLines = (int)(projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
         int numHzLines = (int)(projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
 
-        int width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH + 2;
-        int height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT + 2;
+        float width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH + 2;
+        float height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT + 2;
 
         int maxLines = Math.max(numHzLines, numVtLines);
         Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
